Move Factory namespace to Builder
